# libs/proto/openapi.yaml
openapi: 3.1.0
info:
  title: Duruzi Serve API
  version: 0.1.0
  description: >
    Duruzi Network Serve API for inference, model registry, endpoints, and agent coordination.
    Authentication is via API keys (Bearer tokens). All requests should include an idempotency key
    when creating resources.
servers:
  - url: https://api.duruzi.local
    description: Dev
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APIKey
  parameters:
    RequestId:
      in: header
      name: X-Request-ID
      required: false
      schema: { type: string }
      description: Optional caller-provided request ID for tracing.
  headers:
    RateLimit-Limit:
      description: The number of requests allowed in the current period.
      schema: { type: integer }
    RateLimit-Remaining:
      description: The number of requests remaining in the current period.
      schema: { type: integer }
    RateLimit-Reset:
      description: Seconds until the rate limit resets.
      schema: { type: integer }
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, description: Machine-readable error code }
        message: { type: string, description: Human-friendly message }
        request_id: { type: string }
        details: { type: object, additionalProperties: true }
    InferRequest:
      type: object
      required: [input]
      properties:
        input: { type: string }
        params:
          type: object
          additionalProperties: [string, number, boolean]
    InferResponse:
      type: object
      required: [output, usage]
      properties:
        output: { type: string }
        endpoint_id: { type: string }
        usage:
          type: object
          required: [prompt_tokens, completion_tokens]
          properties:
            prompt_tokens: { type: integer }
            completion_tokens: { type: integer }
...
security:
  - BearerAuth: []
paths:
...
  /v1/infer/{endpoint_id}:
    post:
      summary: Run inference on a deployed endpoint
      parameters:
        - name: endpoint_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InferRequest' }
      responses:
        '200':
          description: Inference completed
          headers:
            RateLimit-Limit: { $ref: '#/components/headers/RateLimit-Limit' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimit-Remaining' }
            RateLimit-Reset: { $ref: '#/components/headers/RateLimit-Reset' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InferResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Rate limited
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...